Test: match API test FS structure to impl
In support of easing the task of locating where to discover or add tests, the filesystem structure for the tests should match that of the implementation.

Note that this change introduced some sort of test pollution. That is resolved via additional cleanup performed `on_exit` as specified in `PlaywrightTest.Case`.
