--- test/api/browser_context/storage_state_test.exs
@@ -1,7 +1,0 @@
-defmodule Playwright.BrowserContext.StorageStateTest do
-  use Playwright.TestCase, async: true
-
-  # test_should_capture_local_storage
-  # test_should_set_local_storage
-  # test_should_round_trip_through_the_file
-end
