--- test/api/browser_type/connect_cdp_test.exs
@@ -1,7 +1,0 @@
-defmodule Playwright.BrowserType.ConnectCDPTest do
-  use Playwright.TestCase, async: true
-
-  # test_connect_to_an_existing_cdp_session
-  # test_connect_to_an_existing_cdp_session_twice
-  # test_conect_over_a_ws_endpoint
-end
