--- test/api/browser_type/connect_test.exs
@@ -1,15 +1,0 @@
-defmodule Playwright.BrowserType.ConnectTest do
-  use Playwright.TestCase, async: true
-
-  # test_browser_type_connect_should_be_able_to_reconnect_to_a_browser
-  # test_browser_type_connect_should_be_able_to_connect_two_browsers_at_the_same_time
-  # test_browser_type_connect_disconnected_event_should_be_emitted_when_browser_is_closed_or_server_is_closed
-  # test_browser_type_connect_disconnected_event_should_be_emitted_when_remote_killed_connection
-  # test_browser_type_disconnected_event_should_have_browser_as_argument
-  # test_browser_type_connect_set_browser_connected_state
-  # test_browser_type_connect_should_throw_when_used_after_is_connected_returns_false
-  # test_browser_type_connect_should_reject_navigation_when_browser_closes
-  # test_should_not_allow_getting_the_path
-  # test_prevent_getting_video_path
-  # test_connect_to_closed_server_without_hangs
-end
