--- test/api/check_test.exs
@@ -1,9 +1,0 @@
-# test_check_the_box(page):
-# test_not_check_the_checked_box(page):
-# test_uncheck_the_box(page):
-# test_not_uncheck_the_unchecked_box(page):
-# test_check_the_box_by_label(page):
-# test_check_the_box_outside_label(page):
-# test_check_the_box_inside_label_without_id(page):
-# test_check_radio(page):
-# test_check_the_box_by_aria_role(page):
