--- test/api/launcher_test.exs
@@ -1,9 +1,0 @@
-# test_browser_type_launch_should_reject_all_promises_when_browser_is_closed
-# test_browser_type_launch_should_throw_if_page_argument_is_passed
-# test_browser_type_launch_should_reject_if_launched_browser_fails_immediately
-# test_browser_type_launch_should_reject_if_executable_path_is_invalid
-# test_browser_type_executable_path_should_work
-# test_browser_type_name_should_work
-# test_browser_close_should_fire_close_event_for_all_contexts
-# test_browser_close_should_be_callable_twice
-# test_browser_launch_should_return_background_pages
